refactoring dummy event service, DRY principal; adding bulk_download_event to existing invoker tests
