+++ tests/app/services/download/test_download_queue.py
@@ -2,19 +2,17 @@
 import re
 import time
 from pathlib import Path
-from typing import Any, Dict, List
 
 import pytest
-from pydantic import BaseModel
 from pydantic.networks import AnyHttpUrl
 from requests.sessions import Session
 from requests_testadapter import TestAdapter, TestSession
 
 from invokeai.app.services.download import DownloadJob, DownloadJobStatus, DownloadQueueService
-from invokeai.app.services.events.events_base import EventServiceBase
+from tests.fixtures.event_service import DummyEventService
 
 # Prevent pytest deprecation warnings
-TestAdapter.__test__ = False
+TestAdapter.__test__ = False  # type: ignore
 
 
 @pytest.fixture
@@ -50,28 +48,6 @@ def session() -> Session:
     sess.mount("http://www.civitai.com/models/broken", TestAdapter(b"Not found", status=404))
 
     return sess
-
-
-class DummyEvent(BaseModel):
-    """Dummy Event to use with Dummy Event service."""
-
-    event_name: str
-    payload: Dict[str, Any]
-
-
-# A dummy event service for testing event issuing
-class DummyEventService(EventServiceBase):
-    """Dummy event service for testing."""
-
-    events: List[DummyEvent]
-
-    def __init__(self) -> None:
-        super().__init__()
-        self.events = []
-
-    def dispatch(self, event_name: str, payload: Any) -> None:
-        """Dispatch an event by appending it to self.events."""
-        self.events.append(DummyEvent(event_name=payload["event"], payload=payload["data"]))
 
 
 def test_basic_queue_download(tmp_path: Path, session: Session) -> None:
