+++ tests/app/services/model_install/test_model_install.py
@@ -10,7 +10,6 @@ from pydantic import ValidationError
 from pydantic.networks import Url
 
 from invokeai.app.services.config import InvokeAIAppConfig
-from invokeai.app.services.events.events_base import EventServiceBase
 from invokeai.app.services.model_install import (
     InstallStatus,
     LocalModelSource,
